chore: split common utils from editor
