+++ src/PolygonEditor.ts
@@ -1,4 +1,5 @@
 import { Point, EditorOptions, PolygonCallback } from './types.js';
+import { getDistance, getClosestPointOnSegment, adjustAlpha } from './utils.js';
 
 export class PolygonEditor {
     private container: HTMLElement;
@@ -115,7 +116,7 @@ export class PolygonEditor {
     private getColor(color: string | string[] | undefined, index: number, fallbackColor?: string): string {
         if (color === undefined || color === '') {
              if (fallbackColor) {
-                 return this.adjustAlpha(fallbackColor, 0.2);
+                 return adjustAlpha(fallbackColor, 0.2);
              }
              return "rgba(0, 0, 0, 0.2)";
         }
@@ -123,31 +124,6 @@ export class PolygonEditor {
             return color[index % color.length];
         }
         return color;
-    }
-
-    private adjustAlpha(color: string, alpha: number): string {
-        try {
-            // Hex
-            if (/^#([A-Fa-f0-9]{3}){1,2}$/.test(color)) {
-                let c: any = color.substring(1).split('');
-                if (c.length === 3) {
-                    c = [c[0], c[0], c[1], c[1], c[2], c[2]];
-                }
-                c = '0x' + c.join('');
-                return 'rgba(' + [(c >> 16) & 255, (c >> 8) & 255, c & 255].join(',') + ',' + alpha + ')';
-            }
-            // RGB/RGBA
-            if (color.startsWith('rgb')) {
-                // simple parsing for rgb/rgba
-                const match = color.match(/(\d+(\.\d+)?)/g);
-                if (match && match.length >= 3) {
-                     return `rgba(${match[0]}, ${match[1]}, ${match[2]}, ${alpha})`;
-                }
-            }
-            return color; 
-        } catch(e) {
-            return color;
-        }
     }
 
     // --- internal implementation ---
@@ -231,40 +207,13 @@ export class PolygonEditor {
         };
     }
 
-    private getDistance(p1: Point, p2: Point): number {
-        return Math.sqrt(Math.pow(p1.x - p2.x, 2) + Math.pow(p1.y - p2.y, 2));
-    }
-
-    private getClosestPointOnSegment(p: Point, p1: Point, p2: Point): Point {
-        const x = p1.x;
-        const y = p1.y;
-        const dx = p2.x - x;
-        const dy = p2.y - y;
-        const dot = dx * dx + dy * dy;
-        let t;
-
-        if (dot > 0) {
-            t = ((p.x - x) * dx + (p.y - y) * dy) / dot;
-        } else {
-            t = -1;
-        }
-
-        if (t < 0) {
-            return p1;
-        } else if (t > 1) {
-            return p2;
-        } else {
-            return { x: x + t * dx, y: y + t * dy };
-        }
-    }
-
     private getHoverPoint(pos: Point): { polyIndex: number, pointIndex: number } | null {
         // defined threshold for grabbing a point
         const threshold = this.options.pointRadius * 2;
         
         // Search in current points
         for (let i = 0; i < this.currentPoints.length; i++) {
-            if (this.getDistance(pos, this.currentPoints[i]) <= threshold) {
+            if (getDistance(pos, this.currentPoints[i]) <= threshold) {
                 return { polyIndex: -1, pointIndex: i };
             }
         }
@@ -273,7 +222,7 @@ export class PolygonEditor {
         for (let i = 0; i < this.polygons.length; i++) {
             const poly = this.polygons[i].points;
             for (let j = 0; j < poly.length; j++) {
-                if (this.getDistance(pos, poly[j]) <= threshold) {
+                if (getDistance(pos, poly[j]) <= threshold) {
                     return { polyIndex: i, pointIndex: j };
                 }
             }
@@ -370,8 +319,8 @@ export class PolygonEditor {
                         const p1 = points[j];
                         const p2 = points[(j + 1) % points.length]; // wrapping
                         
-                        const closest = this.getClosestPointOnSegment(this.mousePos, p1, p2);
-                        const dist = this.getDistance(this.mousePos, closest);
+                        const closest = getClosestPointOnSegment(this.mousePos, p1, p2);
+                        const dist = getDistance(this.mousePos, closest);
                         
                         if (dist <= threshold) {
                             this.ghostPoint = {
@@ -526,7 +475,7 @@ export class PolygonEditor {
 
         // Draw ghost point
         if (this.ghostPoint) {
-            this.ctx.fillStyle = this.adjustAlpha(this.options.pointColor, 0.5);
+            this.ctx.fillStyle = adjustAlpha(this.options.pointColor, 0.5);
             this.ctx.setLineDash([2, 2]);
             this.ctx.beginPath();
             this.ctx.arc(this.ghostPoint.point.x, this.ghostPoint.point.y, this.options.pointRadius, 0, Math.PI * 2);
